refactor brillig_blocks
